move a function set of display options to its own category
