+++ src/components/classifier/ClassificationPanel.js
@@ -8,7 +8,7 @@ import EStyleSheet from 'react-native-extended-stylesheet'
 import PropTypes from 'prop-types';
 
 import FontedText from '../common/FontedText'
-import * as colorModes from '../../actions/colorModes'
+import * as colorModes from '../../displayOptions/colorModes'
 
 class ClassificationPanel extends Component {
   render() {
