+++ src/components/classifier/QuestionClassifier.js
@@ -28,7 +28,7 @@ import {
 import FullScreenImage from '../FullScreenImage'
 import TapableSubject from './TapableSubject';
 
-import * as colorModes from '../../actions/colorModes'
+import * as colorModes from '../../displayOptions/colorModes'
 
 class QuestionClassifier extends Component {
